requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-google-genai==0.0.11
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.1
langchain-xai
